refactor: merge inferred and provided schema
